Graduate SleepGraph2 to SleepGraph, delete unused files
