+++ lib/widgets/sleep_graph/paintable/paintable.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:graphsleepcalc/widgets/sleep_graph/hit_event/hit_event.dart';
 
 // TODO implements HitTestable/HitTestTarget or BoxHitTest class?
 abstract class Paintable {
-
-  // TODO change to a comprehensive hitbox solution
+  // TODO change to e comprehensive hitbox solution
   bool isInteractable;
   Offset offset;
 
-  Paintable () : isInteractable = true;
+  Paintable() : isInteractable = true;
 
   void paint(Canvas canvas, Size size);
 
@@ -17,4 +17,6 @@ abstract class Paintable {
   bool hitTest(GestureEvent hitOffset);
 
   void onGestureEvent(GestureEvent event);
+  void onTapEvent(TapEvent event);
+  void onHorizontalDragEvent(HorizontalDragEvent event);
 }
