+++ lib/widgets/sleep_graph/paintable/sentinel.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:graphsleepcalc/widgets/sleep_graph/chart_data.dart';
+import 'package:graphsleepcalc/widgets/sleep_graph/graph_context.dart';
 import 'package:graphsleepcalc/widgets/sleep_graph/hit_event/hit_event.dart';
 import 'package:graphsleepcalc/widgets/sleep_graph/paintable/paintable.dart';
+import 'package:intl/intl.dart';
 
 abstract class Sentinel extends Paintable {
-
   static const double SENTINEL_DIAMETER = 35.0;
   static const double SENTINEL_RADIUS = SENTINEL_DIAMETER / 2;
   static const double HIT_BOX_MARGIN = 25.0;
+  static const double SENTINEL_BODY_LINE_MARGIN = 25.0;
 
   // What sentinel is current being dragged. Null if no sentinel being dragged.
   // Necessary to account for delays between calls to hitTest() so it doesn't
@@ -25,29 +26,32 @@ abstract class Sentinel extends Paintable {
   get bodyPaint => _bodyPaint;
   set bodyPaint (paint) => _bodyPaint = paint;
 
-  ChartData _chartData;
+  ValueNotifier<DateTime> notifier = ValueNotifier(null);
 
-  Sentinel (this._chartData) : isLocked = false;
+  GraphContext _graphContext;
+
+  Sentinel(this._graphContext) : isLocked = false;
+
+  static String formatTimeString(DateTime time) {
+    String timeString = DateFormat('h:mma').format(time);
+    return timeString.toLowerCase().substring(0, timeString.length - 1);
+  }
 
   /// Paint the body of this sentinel
-  // TODO pass a GridData object so Paintables know where to paint themselves?
   @override
   void paint(Canvas canvas, Size size) {
     _paintSentinelLine(canvas, size);
     canvas.drawCircle(offset, Sentinel.SENTINEL_RADIUS, bodyPaint);
   }
 
-  void _paintSentinelLine(Canvas canvas, Size size) {
-    double dx = offset.dx;
-    double y1 = 125.0; // TODO get line y from sleep graph
-    double y2 = 300; // TODO get bottom y from slee graph
-    // canvas.drawLine(Offset(dx, y1), Offset(dx, y2), _bodyPaint);
-    canvas.drawLine(Offset(dx, y1), Offset(dx, size.height), _bodyPaint);
+  void _paintSentinelLine(Canvas canvas, Size size) {    
+    Offset o1 = Offset(offset.dx, _graphContext.sleepCycleMinY);
+    Offset o2 = Offset(offset.dx, _graphContext.sleepCycleMaxY); // TODO calc y of sleep cycle
+    canvas.drawLine(o1, o2, _bodyPaint);
   }
 
   /// Returns true if [event] is within this Sentinel's hitbox.
   /// If hit, update location to [event.localPosition].
-  // TODO handle case when sentinels are overlapping
   @override
   bool hitTest(GestureEvent event) {
     if (!isInteractable) {
@@ -56,9 +60,8 @@ abstract class Sentinel extends Paintable {
 
     // Check that [event] is within the chart area
     if (event.localPosition != null
-        && (event.localPosition.dx < _chartData.xAxisX + SENTINEL_RADIUS
-        || event.localPosition.dy > _chartData.yAxisY)) {
-      _draggingSentinel = null; // Drop the sentinel
+        && (event.localPosition.dx < _graphContext.sleepCycleMinX
+        || event.localPosition.dy > _graphContext.maxY)) {
       return false;
     }
 
@@ -69,36 +72,49 @@ abstract class Sentinel extends Paintable {
     }
 
     // TODO (x + width - offset.dx).abs()
-    var hitOffset = event.localPosition;
+    Offset hitOffset = event.localPosition + _graphContext.viewPane;
     bool hit = (offset.dx - hitOffset.dx).abs() < HIT_BOX_MARGIN
             && (offset.dy - hitOffset.dy).abs() < HIT_BOX_MARGIN;
 
     return hit;
   }
 
-  @override
-  void onGestureEvent(GestureEvent event);
-
   // TODO change to private method that children can call
+  @override
   void onHorizontalDragEvent(HorizontalDragEvent event) {
-    // print('Sentinel.onHorizontalDragEvent(${event.runtimeType})');
     Offset eventOffset = event.localPosition;
     if (event is HorizontalDragStartEvent) {
       _draggingSentinel = this;
-    } else if (event is HorizontalDragUpdateEvent) {
-      if (this == _draggingSentinel) {
-        offset = Offset(eventOffset.dx, offset.dy);
-      }
+    } else if (event is HorizontalDragUpdateEvent 
+        && this == _draggingSentinel && !isLocked) {
+      setOffset(Offset(eventOffset.dx + _graphContext.viewPane.dx, offset.dy));
     } else if (event is HorizontalDragEndEvent ||
         event is HorizontalDragCancelEvent) {
       _draggingSentinel = null;
     }
   }
+
+  void setOffset(Offset offset) {
+    this.offset = offset;
+    var minutesPerX 
+        = _graphContext.sleepCycleMinutes / _graphContext.sleepCycleWidth;
+    var xDiff = this.offset.dx - _graphContext.sleepCycleMinX;
+    var minutesOffset = Duration(minutes: (minutesPerX * xDiff).toInt());
+    var time = DateTime.now().add(minutesOffset);
+    notifier.value = time;
+  }
+
+  double bodyY(GraphContext graphContext) {
+    return graphContext.sleepCycleMinY - SENTINEL_BODY_LINE_MARGIN;
+  }
+
+  Offset _initialOffset(double numCycles) {
+    Offset sleepCycleStart = _graphContext.sleepCycleToOffset(numCycles);
+    return Offset(sleepCycleStart.dx, bodyY(_graphContext));
+  }
 }
 
-
 class SleepSentinel extends Sentinel {
-
   final _bigCraterRadius = Sentinel.SENTINEL_RADIUS * 0.3;
   final _mediumCraterRadius = Sentinel.SENTINEL_RADIUS * 0.2;
   final _smallCraterRadius = Sentinel.SENTINEL_RADIUS * 0.15;
@@ -107,25 +123,24 @@ class SleepSentinel extends Sentinel {
       ..color = Colors.grey[300]
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.5;
-  
   final Paint _lockedBodyPaint = Paint()
       ..color = Colors.grey[500]
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.5;
-
   final Paint _craterPaint = Paint()
       ..color = Colors.grey[600]
       ..style = PaintingStyle.fill;
-
-  SleepSentinel (_chartData) : super(_chartData) {
+  
+  SleepSentinel(_graphContext) : super(_graphContext) {
     // TODO set bodyPaint by calling super constructor
     bodyPaint = _unlockedBodyPaint;
 
-     // TODO get init location from sleep graph
-    offset = Offset(100, 100);
+    _graphContext.sleepTimeNotifier = notifier;
+
+    setOffset(_initialOffset(0.0));
   }
 
-  /// Have SleepSentinel paint itself at [offset].
+  /// Have SleepSentinel paint itself at [offset]
   @override
   void paint(Canvas canvas, Size size) {
     super.paint(canvas, size);
@@ -144,34 +159,48 @@ class SleepSentinel extends Sentinel {
 
   @override
   void onGestureEvent(GestureEvent event) {
-    // print('SleepSentinel.onGestureEvent(${event.runtimeType})');
     if (event is TapUpEvent) {
-      isLocked = !isLocked;
-      bodyPaint = isLocked ? _lockedBodyPaint : _unlockedBodyPaint;
+      onTapEvent(event);
     } else if (event is HorizontalDragEvent) {
       onHorizontalDragEvent(event);
     }
   }
+
+  @override
+  void onTapEvent(TapEvent event) {
+    isLocked = !isLocked;
+    bodyPaint = isLocked ? _lockedBodyPaint : _unlockedBodyPaint;
+  }
+
+  @override
+  void setOffset(Offset offset) {
+    if (_graphContext.wakeSentinelX != null 
+        && _graphContext.wakeSentinelX <= offset.dx) {
+      return;
+    }
+
+    super.setOffset(offset);
+    _graphContext.sleepSentinelX = offset.dx;
+  }
 }
 
-
 class WakeSentinel extends Sentinel {
-
   final Paint _unlockedBodyPaint = Paint()
       ..color = Colors.yellow[600]
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.5;
-
   final Paint _lockedBodyPaint = Paint()
       ..color = Colors.yellow[800]
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.5;
-
-  WakeSentinel (_chartData) : super(_chartData) {
+  
+  WakeSentinel(_graphContext) : super(_graphContext) {
     // TODO set bodyPaint by calling super constructor
     super.bodyPaint = _unlockedBodyPaint;
 
-    offset = Offset(300, 100); // TODO get init location from sleep graph
+    _graphContext.wakeTimeNotifier = notifier;
+
+    setOffset(_initialOffset(3.0));
   }
 
   /// Have WakeSentinel paint itself at [offset].
@@ -183,12 +212,27 @@ class WakeSentinel extends Sentinel {
 
   @override
   void onGestureEvent(GestureEvent event) {
-    // print('WakeSentinel.onGestureEvent(${event.runtimeType})');
     if (event is TapUpEvent) {
-      isLocked = !isLocked;
-      bodyPaint = isLocked ? _lockedBodyPaint : _unlockedBodyPaint;
+      onTapEvent(event);
     } else if (event is HorizontalDragEvent) {
       onHorizontalDragEvent(event);
     }
+  }
+
+  @override
+  void onTapEvent(TapEvent event) {
+    isLocked = !isLocked;
+    bodyPaint = isLocked ? _lockedBodyPaint : _unlockedBodyPaint;
+  }
+
+  @override
+  void setOffset(Offset offset) {
+    if (_graphContext.sleepSentinelX != null 
+        && _graphContext.sleepSentinelX >= offset.dx) {
+      return;
+    }
+
+    super.setOffset(offset);
+    _graphContext.wakeSentinelX = offset.dx;
   }
 }
